+++ fabric/src/main/java/io/github/sakurawald/fuji/module/mixin/world/keep_inventory_rule/ServerPlayerEntityMixin.java
@@ -1,0 +1,18 @@
+package io.github.sakurawald.fuji.module.mixin.world.keep_inventory_rule;
+
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import com.llamalad7.mixinextras.sugar.Local;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.world.GameRules;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(ServerPlayerEntity.class)
+public class ServerPlayerEntityMixin {
+
+    @ModifyExpressionValue(method = "copyFrom", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/GameRules;getBoolean(Lnet/minecraft/world/GameRules$Key;)Z"))
+    boolean f(boolean original, @Local(argsOnly = true) ServerPlayerEntity oldPlayer) {
+        // NOTE: For `keepInventory` game rule. Its value is checked in copyFrom() method after the player is dead. The value comes from the re-spawn dimension's world properties.
+        return oldPlayer.getWorld().getGameRules().getBoolean(GameRules.KEEP_INVENTORY);
+    }
+}