+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - feature: improve the compatibility with other mods.
   - feature: support `per-dimension difficulty`.
   - feature: support `per-dimension time of day`. (Now you can `sleep` and set `spawn point` in `extra dimensions`.)
+  - feature: support `per-dimension gamerules`.
   - feature: support `per-dimension weather`.
 - [fuji] feature: add `click promopt` for `/fuji inspect configurations` command.
 - [core] feature: improve the performance of search function.